organization; move useContacts and enable contacts modal to teams/common
