+++ shared/people/invite-friends/invite-contacts.native.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react'
 import * as Kb from '../../common-adapters/mobile.native'
 import {Section as _Section} from '../../common-adapters/section-list'
-import useContacts from '../../teams/invite-by-contact/use-contacts.native'
+import useContacts from '../../teams/common/use-contacts.native'
 import {Contact} from '../../teams/invite-by-contact/index.native'
 import {memoize} from '../../util/memoize'
 import * as Container from '../../util/container'
