+++ shared/teams/add-members-wizard/index.stories.tsx
@@ -3,7 +3,6 @@ import * as Sb from '../../stories/storybook'
 import * as Container from '../../util/container'
 import * as Constants from '../../constants/teams'
 import AddFromWhere from './add-from-where'
-import EnableContacts from './enable-contacts'
 import AddEmail from './add-email'
 import AddPhone from './add-phone'
 import AddMembersConfirm from './confirm'
@@ -64,7 +63,6 @@ const load = () => {
     .addDecorator(story => <Sb.MockStore store={commonStore}>{story()}</Sb.MockStore>)
     .add('Add from where', () => <AddFromWhere {...Sb.createNavigator({})} />)
     .add('Add from where (new team)', () => <AddFromWhere {...fromWhereNewProps} />)
-    .add('Enable contacts', () => <EnableContacts onClose={Sb.action('onClose')} />)
     .add('Add by email', () => <AddEmail teamID={fakeTeamID} errorMessage="" />)
     .add('Add by phone', () => <AddPhone />)
