+++ shared/teams/invite-by-contact/team-invite-by-contacts.native.tsx
@@ -4,7 +4,7 @@ import * as React from 'react'
 import * as SettingsConstants from '../../constants/settings'
 import * as TeamsGen from '../../actions/teams-gen'
 import * as Types from '../../constants/types/teams'
-import useContacts from './use-contacts.native'
+import useContacts from '../common/use-contacts.native'
 import {Contact, ContactRowProps, InviteByContact} from './index.native'
 
 // Seitan invite names (labels) look like this: "[name] ([phone number])". Try
